+++ Sources/KeyboardKit/_Migrations/DictationMigrations.swift
@@ -12,3 +12,15 @@ public extension Dictation {
     @available(*, deprecated, renamed: "StandardService", message: "Migration Deprecation, will be removed in 9.1!")
     typealias ProKeyboardService = StandardService
 }
+
+public extension View {
+
+    @available(*, deprecated, renamed: "StandardService", message: "Migration Deprecation, will be removed in 9.1! You must now pass in a keyboardContext and an openURL for dictation to work.")
+    func keyboardDictation<Overlay: View>(
+        context: DictationContext,
+        speechRecognizer: DictationSpeechRecognizer,
+        @ViewBuilder overlay: () -> Overlay
+    ) -> some View {
+        self
+    }
+}